NFC: Fix a narrowing conversion from size_t to int64_t when constructing a VectorType.
PiperOrigin-RevId: 253125435

+++ unittests/IR/AttributeTest.cpp
@@ -60,7 +60,7 @@ TEST(DenseSplatTest, BoolSplat) {
 }
 
 TEST(DenseSplatTest, LargeBoolSplat) {
-  constexpr size_t boolCount = 56;
+  constexpr int64_t boolCount = 56;
 
   MLIRContext context;
   IntegerType boolTy = IntegerType::get(1, &context);
